+++ src/pages/Dashboard/Automations/AddAction.vue
@@ -130,9 +130,21 @@ export default {
     },
     disableNext() {
       if (this.step.name === 'openToConfig') {
+        if (
+          this.isPagerDuty &&
+          this.apiToken &&
+          this.routingKey &&
+          this.severity
+        )
+          return false
         if (!this.actionConfigArray.length) return true
         else return false
       } else {
+        if (this.step.name === 'addTwilioConfig') {
+          if (!!this.authToken && !!this.messagingService && !!this.accountSid)
+            return false
+          return true
+        }
         return false
       }
     },
@@ -172,18 +184,24 @@ export default {
     to() {
       const configTo =
         this.actionConfigArray.length > 0
-          ? this.actionConfigArray.toString()
-          : this.secretName || ''
+          ? `to ${this.actionConfigArray.toString()}`
+          : this.secretName
+          ? `to ${this.secretName}`
+          : ''
+      const config =
+        !this.isTwilio && this.actionType?.config?.to
+          ? `to ${this.actionType?.config?.to}`
+          : this.isTwilio && this.actionType?.config?.to.length
+          ? this.actionType?.config?.to.toString()
+          : false
       return this.actionType?.type === 'EMAIL'
-        ? configTo || this.actionType?.config?.to || 'to this email address.'
+        ? configTo || config || 'to this email address.'
         : this.actionType.type === 'WEBHOOK'
-        ? this.actionType?.config?.to || 'to this web address.'
+        ? config || 'to this web address.'
         : this.actionType?.type === 'TWILIO'
-        ? configTo ||
-          this.actionType?.config?.to?.toString() ||
-          'to this phone number'
+        ? configTo || config || 'to this phone number'
         : this.actionType?.type === 'SLACK_WEBHOOK'
-        ? configTo || this.actionType?.config?.url || 'here.'
+        ? configTo || config || 'here.'
         : this.actionType.type === 'PAGERDUTY'
         ? 'using this config.'
         : this.actionType.type === 'MS_TEAMS'