+++ Changelog.md
@@ -8,7 +8,18 @@
 
 ### Bugfixes
 
+<<<<<<< HEAD
 - Fix issues with pre-populating and resetting job templates and task definitions for run configs - [#839](https://github.com/PrefectHQ/ui/pull/839)
+=======
+- Check for the correct values when disabling next for the Pager Duty action - [#835](https://github.com/PrefectHQ/ui/pull/835)
+
+## 2021-05-17
+
+### Bugfixes
+
+- Covert this.allowedUsers in the Members component to a computed property to use the nullish coalescing operator/check for total allowed users -[#841](https://github.com/PrefectHQ/ui/pull/841)
+- Beautify json and add json type to default parameters json edit option to stop value flicker - [#837](https://github.com/PrefectHQ/ui/pull/837)
+>>>>>>> 7e68094dbf76510039632e0317a6d9c1ee435f8b
 
 ## 2021-05-14
 
@@ -19,6 +30,7 @@
 - Add new states to the license store - [#763](https://github.com/PrefectHQ/ui/pull/763)
 - Add a tooltip for task run names in the flow run task table - [#814](https://github.com/PrefectHQ/ui/pull/814)
 - Add "Task" to the "Runs Today" to make it clear it's "Task runs today" - [#816](https://github.com/PrefectHQ/ui/pull/816)
+- Add cancel all and work queue paused system actions, VU for clear late runs on the dashboard - [#825](https://github.com/PrefectHQ/ui/pull/825)
 
 ### Bugfixes